Moved definition of `randomData()` method to base test class `tests/PHPJsonUtilitiesTest.php`.
